Forward declare more expressions.
