+++ src/include/binder/bind_node_visitor.h
@@ -2,21 +2,16 @@
 
 #include <memory>
 #include <string>
-#include <utility>
 #include <vector>
 
-#include "binder/binder_sherpa.h"
 #include "binder/sql_node_visitor.h"
-#include "catalog/catalog_accessor.h"
 #include "catalog/catalog_defs.h"
-#include "parser/expression/column_value_expression.h"
-#include "parser/postgresparser.h"
-#include "parser/select_statement.h"
 #include "type/type_id.h"
 
 namespace noisepage {
 
 namespace parser {
+class AbstractExpression;
 class AggregateExpression;
 class CaseExpression;
 class ConstantValueExpression;
@@ -34,6 +29,7 @@ class CatalogAccessor;
 
 namespace binder {
 class BinderContext;
+class BinderSherpa;
 
 /**
  * Interface to be notified of the composition of a bind node.
@@ -46,6 +42,9 @@ class BindNodeVisitor : public SqlNodeVisitor {
    * @param db_oid oid of the connected database
    */
   BindNodeVisitor(common::ManagedPointer<catalog::CatalogAccessor> catalog_accessor, catalog::db_oid_t db_oid);
+
+  /** Destructor. Must be defined due to forward declaration. */
+  ~BindNodeVisitor();
 
   /**
    * Perform binding on the passed in tree. Bind the relation names to oids
@@ -96,7 +95,7 @@ class BindNodeVisitor : public SqlNodeVisitor {
 
  private:
   /** BinderSherpa which stores metadata (e.g. type information) across Visit calls. **/
-  std::unique_ptr<BinderSherpa> sherpa_ = nullptr;
+  std::unique_ptr<BinderSherpa> sherpa_;
   /** Current context of the query or subquery */
   common::ManagedPointer<BinderContext> context_ = nullptr;
 
@@ -104,9 +103,7 @@ class BindNodeVisitor : public SqlNodeVisitor {
   const common::ManagedPointer<catalog::CatalogAccessor> catalog_accessor_;
   const catalog::db_oid_t db_oid_;
 
-  static void InitTableRef(const common::ManagedPointer<parser::TableRef> node) {
-    if (node->table_info_ == nullptr) node->table_info_ = std::make_unique<parser::TableInfo>();
-  }
+  static void InitTableRef(const common::ManagedPointer<parser::TableRef> node);
 
   /**
    * Change the type of exprs_ of order_by_description from ConstantValueExpression to ColumnValueExpression.
@@ -116,17 +113,7 @@ class BindNodeVisitor : public SqlNodeVisitor {
   void UnifyOrderByExpression(common::ManagedPointer<parser::OrderByDescription> order_by_description,
                               const std::vector<common::ManagedPointer<parser::AbstractExpression>> &select_items);
 
-  void ValidateDatabaseName(const std::string &db_name) {
-    if (!(db_name.empty())) {
-      const auto db_oid = catalog_accessor_->GetDatabaseOid(db_name);
-      if (db_oid == catalog::INVALID_DATABASE_OID)
-        throw BINDER_EXCEPTION(fmt::format("Database \"{}\" does not exist", db_name),
-                               common::ErrorCode::ERRCODE_UNDEFINED_DATABASE);
-      if (db_oid != db_oid_)
-        throw BINDER_EXCEPTION("cross-database references are not implemented: ",
-                               common::ErrorCode::ERRCODE_FEATURE_NOT_SUPPORTED);
-    }
-  }
+  void ValidateDatabaseName(const std::string &db_name);
 };
 
 }  // namespace binder
