+++ src/include/common/hash_defs.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <cstdint>
+
+namespace noisepage::common {
+/**
+ * This is our typedef that we use throughout the entire code to represent a hash value.
+ */
+using hash_t = uint64_t;
+}  // namespace noisepage
