+++ src/include/parser/expression/case_expression.h
@@ -38,12 +38,7 @@ class CaseExpression : public AbstractExpression {
      * Hash the current WhenClause.
      * @return hash of WhenClause
      */
-    common::hash_t Hash() const {
-      common::hash_t hash = condition_->Hash();
-      hash = common::HashUtil::CombineHashes(hash, condition_->Hash());
-      hash = common::HashUtil::CombineHashes(hash, then_->Hash());
-      return hash;
-    }
+    common::hash_t Hash() const;
 
     /**
      * Derived expressions should call this base method
@@ -130,7 +125,7 @@ class CaseExpression : public AbstractExpression {
   /** @return default clause, if it exists */
   common::ManagedPointer<AbstractExpression> GetDefaultClause() const { return common::ManagedPointer(default_expr_); }
 
-  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override { v->Visit(common::ManagedPointer(this)); }
+  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override;
 
   /** @return expression serialized to json */
   nlohmann::json ToJson() const override;
