+++ src/include/parser/expression/column_value_expression.h
@@ -152,7 +152,7 @@ class ColumnValueExpression : public AbstractExpression {
    */
   void DeriveExpressionName() override;
 
-  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override { v->Visit(common::ManagedPointer(this)); }
+  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override;
 
   /**
    * @return expression serialized to json
