+++ src/include/parser/expression/constant_value_expression.h
@@ -227,7 +227,7 @@ class ConstantValueExpression : public AbstractExpression {
   template <typename T>
   T Peek() const;
 
-  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override { v->Visit(common::ManagedPointer(this)); }
+  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override;
 
   /** @return A string representation of this ConstantValueExpression. */
   std::string ToString() const;
