+++ src/include/parser/expression/derived_value_expression.h
@@ -56,7 +56,7 @@ class DerivedValueExpression : public AbstractExpression {
 
   bool operator==(const AbstractExpression &rhs) const override;
 
-  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override { v->Visit(common::ManagedPointer(this)); }
+  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override;
 
   /** @return expression serialized to json */
   nlohmann::json ToJson() const override;
