+++ src/include/parser/expression/operator_expression.h
@@ -42,7 +42,7 @@ class OperatorExpression : public AbstractExpression {
 
   void DeriveReturnValueType() override;
 
-  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override { v->Visit(common::ManagedPointer(this)); }
+  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override;
 };
 
 DEFINE_JSON_HEADER_DECLARATIONS(OperatorExpression);
