+++ src/include/parser/expression/star_expression.h
@@ -36,7 +36,7 @@ class StarExpression : public AbstractExpression {
     return Copy();
   }
 
-  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override { v->Visit(common::ManagedPointer(this)); }
+  void Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) override;
 };
 
 DEFINE_JSON_HEADER_DECLARATIONS(StarExpression);
