+++ src/parser/expression/column_value_expression.cpp
@@ -1,5 +1,7 @@
 #include "parser/expression/column_value_expression.h"
 
+#include "binder/sql_node_visitor.h"
+#include "common/hash_util.h"
 #include "common/json.h"
 
 namespace noisepage::parser {
@@ -43,6 +45,10 @@ void ColumnValueExpression::DeriveExpressionName() {
     this->SetExpressionName(this->GetAlias());
   else
     this->SetExpressionName(column_name_);
+}
+
+void ColumnValueExpression::Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) {
+  v->Visit(common::ManagedPointer(this));
 }
 
 nlohmann::json ColumnValueExpression::ToJson() const {
