+++ src/parser/expression/default_value_expression.cpp
@@ -1,5 +1,6 @@
 #include "parser/expression/default_value_expression.h"
 
+#include "binder/sql_node_visitor.h"
 #include "common/json.h"
 
 namespace noisepage::parser {
@@ -8,6 +9,10 @@ std::unique_ptr<AbstractExpression> DefaultValueExpression::Copy() const {
   auto expr = std::make_unique<DefaultValueExpression>();
   expr->SetMutableStateForCopy(*this);
   return expr;
+}
+
+void DefaultValueExpression::Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) {
+  v->Visit(common::ManagedPointer(this));
 }
 
 DEFINE_JSON_BODY_DECLARATIONS(DefaultValueExpression);
