+++ src/parser/expression/function_expression.cpp
@@ -1,5 +1,7 @@
 #include "parser/expression/function_expression.h"
 
+#include "binder/sql_node_visitor.h"
+#include "common/hash_util.h"
 #include "common/json.h"
 
 namespace noisepage::parser {
@@ -33,6 +35,16 @@ std::vector<std::unique_ptr<AbstractExpression>> FunctionExpression::FromJson(co
   exprs.insert(exprs.end(), std::make_move_iterator(e1.begin()), std::make_move_iterator(e1.end()));
   func_name_ = j.at("func_name").get<std::string>();
   return exprs;
+}
+
+void FunctionExpression::Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) {
+  v->Visit(common::ManagedPointer(this));
+}
+
+common::hash_t FunctionExpression::Hash() const {
+  common::hash_t hash = AbstractExpression::Hash();
+  hash = common::HashUtil::CombineHashes(hash, common::HashUtil::Hash(func_name_));
+  return hash;
 }
 
 DEFINE_JSON_BODY_DECLARATIONS(FunctionExpression);
