+++ src/parser/expression/operator_expression.cpp
@@ -1,5 +1,6 @@
 #include "parser/expression/operator_expression.h"
 
+#include "binder/sql_node_visitor.h"
 #include "common/json.h"
 
 namespace noisepage::parser {
@@ -36,6 +37,10 @@ void OperatorExpression::DeriveReturnValueType() {
   const auto &type = (*max_type_child)->GetReturnValueType();
   NOISEPAGE_ASSERT(type <= type::TypeId::DECIMAL, "Invalid operand type in Operator Expression.");
   this->SetReturnValueType(type);
+}
+
+void OperatorExpression::Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) {
+  v->Visit(common::ManagedPointer(this));
 }
 
 DEFINE_JSON_BODY_DECLARATIONS(OperatorExpression);
