+++ src/parser/expression/star_expression.cpp
@@ -1,5 +1,6 @@
 #include "parser/expression/star_expression.h"
 
+#include "binder/sql_node_visitor.h"
 #include "common/json.h"
 
 namespace noisepage::parser {
@@ -8,6 +9,10 @@ std::unique_ptr<AbstractExpression> StarExpression::Copy() const {
   auto expr = std::make_unique<StarExpression>();
   expr->SetMutableStateForCopy(*this);
   return expr;
+}
+
+void StarExpression::Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) {
+  v->Visit(common::ManagedPointer(this));
 }
 
 DEFINE_JSON_BODY_DECLARATIONS(StarExpression);
