+++ src/parser/expression/table_star_expression.cpp
@@ -1,5 +1,6 @@
 #include "parser/expression/table_star_expression.h"
 
+#include "binder/sql_node_visitor.h"
 #include "common/json.h"
 
 namespace noisepage::parser {
@@ -10,6 +11,10 @@ std::unique_ptr<AbstractExpression> TableStarExpression::Copy() const {
   expr->target_table_specified_ = target_table_specified_;
   expr->target_table_ = target_table_;
   return expr;
+}
+
+void TableStarExpression::Accept(common::ManagedPointer<binder::SqlNodeVisitor> v) {
+  v->Visit(common::ManagedPointer(this));
 }
 
 DEFINE_JSON_BODY_DECLARATIONS(TableStarExpression);
